+++ src/test/java/com/daangn/dangunmarket/domain/post/postlike/PostLikeServiceTest.java
@@ -3,6 +3,7 @@ package com.daangn.dangunmarket.domain.post.postlike;
 import com.daangn.dangunmarket.domain.member.model.Member;
 import com.daangn.dangunmarket.domain.member.model.NickName;
 import com.daangn.dangunmarket.domain.member.repository.MemberJpaRepository;
+import com.daangn.dangunmarket.domain.post.exception.InvalidPostLikeException;
 import com.daangn.dangunmarket.domain.post.model.Category;
 import com.daangn.dangunmarket.domain.post.model.LocationPreference;
 import com.daangn.dangunmarket.domain.post.model.Post;
@@ -15,7 +16,6 @@ import com.daangn.dangunmarket.domain.post.repository.post.PostRepository;
 import com.daangn.dangunmarket.domain.post.service.PostLikeService;
 import com.daangn.dangunmarket.global.GeometryTypeFactory;
 import com.daangn.dangunmarket.global.exception.EntityNotFoundException;
-import com.daangn.dangunmarket.global.exception.InvalidPostLikeException;
 import com.daangn.dangunmarket.global.response.ErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.assertj.core.api.Assertions;
